+++ Modules/ADUtils/UITableViewHeaderFooterView+PreferredLayoutHeight.swift
@@ -16,24 +16,6 @@ extension UITableViewHeaderFooterView {
      - parameter fittingWidth: The biggest width the view can get
      */
     public func ad_preferredContentViewLayoutHeight(fittingWidth: CGFloat) -> CGFloat {
-        var bounds = self.bounds
-        bounds.size.width = fittingWidth
-        self.bounds = bounds
-        let views = ["contentView": contentView]
-        let visualFormats = ["H:|[contentView]|", "V:|[contentView]|"]
-        let constraints = visualFormats
-            .map {
-                NSLayoutConstraint.constraints(
-                    withVisualFormat: $0,
-                    options: [],
-                    metrics: nil,
-                    views: views
-                )
-            }
-            .flatMap { $0 }
-        addConstraints(constraints)
-        let height = contentView.ad_preferredLayoutHeight(fittingWidth: fittingWidth)
-        removeConstraints(constraints)
-        return height
+        return contentView.ad_preferredLayoutHeight(fittingWidth: fittingWidth)
     }
 }